requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
pytz==2024.1
tzlocal==5.2
cachetools==5.3.3
tornado==6.4
certifi==2024.2.2
requests==2.32.3
